Extract domains, tags, and misc into subfiles
