+++ src/services/domains/domains.schema.ts
@@ -1,0 +1,6 @@
+import { object, string } from 'yup';
+
+export const importZoneSchema = object({
+  domain: string().required(),
+  remote_nameserver: string().required(),
+});
