+++ src/services/domains/index.ts
@@ -1,0 +1,17 @@
+export {
+  getDomains,
+  getDomains$,
+  getDomain,
+  createDomain,
+  updateDomain,
+  deleteDomain,
+  cloneDomain,
+  importZone,
+} from './domains';
+
+export {
+  getDomainRecords,
+  createDomainRecord,
+  updateDomainRecord,
+  deleteDomainRecord,
+} from './records';
