+++ src/services/misc/index.ts
@@ -1,0 +1,3 @@
+export {
+  getRegions
+} from './misc';
