+++ src/services/networking/index.ts
@@ -1,0 +1,6 @@
+export {
+  updateIP,
+  assignAddresses,
+  shareAddresses,
+  getIPs,
+} from './networking';
