+++ src/services/tags/index.ts
@@ -1,0 +1,7 @@
+export {
+  getTags,
+  createTag,
+  deleteTag,
+  Tag,
+  TagRequest,
+} from './tags';
